requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
schedule==1.2.1
python-dotenv==1.0.1
PyYAML==6.0.1
reportlab==4.1.0
python-docx==1.1.0
Flask==3.0.2
click==8.1.7
tabulate==0.9.0
